requirements.txt
torch
torchvision
transformers
pillow
tqdm
numpy
matplotlib
